move functions to a new file
