+++ internal/sidekick/swift/codec.go
@@ -17,9 +17,7 @@ package swift
 import (
 	"fmt"
 	"path/filepath"
-	"slices"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/googleapis/librarian/internal/config"
@@ -216,30 +214,4 @@ func (c *codec) addDependency(dep *Dependency) (*Dependency, error) {
 		ann.DependsOn[dep.Name] = dep
 	}
 	return dep, nil
-}
-
-func (c *codec) protoMessageTypeName(m *api.Message) string {
-	var path []string
-	curr := m
-	for curr != nil {
-		path = append(path, pascalCase(curr.Name))
-		curr = curr.Parent
-	}
-	slices.Reverse(path)
-	fullName := fmt.Sprintf("%s%s", ProtoPackagePrefix(m.Package), strings.Join(path, "."))
-	if c.ModulePath != "" {
-		return fmt.Sprintf("%s.%s", c.ModulePath, fullName)
-	}
-	return fullName
-}
-
-func (c *codec) protoEnumTypeName(e *api.Enum) string {
-	if e.Parent == nil {
-		fullName := fmt.Sprintf("%s%s", ProtoPackagePrefix(e.Package), pascalCase(e.Name))
-		if c.ModulePath != "" {
-			return fmt.Sprintf("%s.%s", c.ModulePath, fullName)
-		}
-		return fullName
-	}
-	return fmt.Sprintf("%s.%s", c.protoMessageTypeName(e.Parent), pascalCase(e.Name))
 }
