+++ internal/sidekick/swift/codec_test.go
@@ -266,90 +266,3 @@ func makeRequiredServicesTestModel() *api.API {
 	api.CrossReference(model)
 	return model
 }
-
-func TestProtoMessageAndEnumTypeName(t *testing.T) {
-	parentMsg := &api.Message{
-		Name:    "OuterMessage",
-		ID:      ".test.OuterMessage",
-		Package: "test",
-	}
-	nestedMsg := &api.Message{
-		Name:    "InnerMessage",
-		ID:      ".test.OuterMessage.InnerMessage",
-		Package: "test",
-		Parent:  parentMsg,
-	}
-	topEnum := &api.Enum{
-		Name:    "TopEnum",
-		ID:      ".test.TopEnum",
-		Package: "test",
-	}
-	nestedEnum := &api.Enum{
-		Name:    "NestedEnum",
-		ID:      ".test.OuterMessage.NestedEnum",
-		Package: "test",
-		Parent:  parentMsg,
-	}
-
-	model := api.NewTestAPI([]*api.Message{parentMsg, nestedMsg}, []*api.Enum{topEnum, nestedEnum}, []*api.Service{})
-	model.PackageName = "test"
-
-	t.Run("with empty ModulePath", func(t *testing.T) {
-		codec := newTestCodec(t, model, map[string]string{})
-		
-		gotMsg := codec.protoMessageTypeName(parentMsg)
-		wantMsg := "Test_OuterMessage"
-		if gotMsg != wantMsg {
-			t.Errorf("protoMessageTypeName(parentMsg) = %q, want %q", gotMsg, wantMsg)
-		}
-
-		gotNestedMsg := codec.protoMessageTypeName(nestedMsg)
-		wantNestedMsg := "Test_OuterMessage.InnerMessage"
-		if gotNestedMsg != wantNestedMsg {
-			t.Errorf("protoMessageTypeName(nestedMsg) = %q, want %q", gotNestedMsg, wantNestedMsg)
-		}
-
-		gotEnum := codec.protoEnumTypeName(topEnum)
-		wantEnum := "Test_TopEnum"
-		if gotEnum != wantEnum {
-			t.Errorf("protoEnumTypeName(topEnum) = %q, want %q", gotEnum, wantEnum)
-		}
-
-		gotNestedEnum := codec.protoEnumTypeName(nestedEnum)
-		wantNestedEnum := "Test_OuterMessage.NestedEnum"
-		if gotNestedEnum != wantNestedEnum {
-			t.Errorf("protoEnumTypeName(nestedEnum) = %q, want %q", gotNestedEnum, wantNestedEnum)
-		}
-	})
-
-	t.Run("with populated ModulePath", func(t *testing.T) {
-		codec := newTestCodec(t, model, map[string]string{
-			"module-path": "TestProtos",
-		})
-
-		gotMsg := codec.protoMessageTypeName(parentMsg)
-		wantMsg := "TestProtos.Test_OuterMessage"
-		if gotMsg != wantMsg {
-			t.Errorf("protoMessageTypeName(parentMsg) = %q, want %q", gotMsg, wantMsg)
-		}
-
-		gotNestedMsg := codec.protoMessageTypeName(nestedMsg)
-		wantNestedMsg := "TestProtos.Test_OuterMessage.InnerMessage"
-		if gotNestedMsg != wantNestedMsg {
-			t.Errorf("protoMessageTypeName(nestedMsg) = %q, want %q", gotNestedMsg, wantNestedMsg)
-		}
-
-		gotEnum := codec.protoEnumTypeName(topEnum)
-		wantEnum := "TestProtos.Test_TopEnum"
-		if gotEnum != wantEnum {
-			t.Errorf("protoEnumTypeName(topEnum) = %q, want %q", gotEnum, wantEnum)
-		}
-
-		gotNestedEnum := codec.protoEnumTypeName(nestedEnum)
-		wantNestedEnum := "TestProtos.Test_OuterMessage.NestedEnum"
-		if gotNestedEnum != wantNestedEnum {
-			t.Errorf("protoEnumTypeName(nestedEnum) = %q, want %q", gotNestedEnum, wantNestedEnum)
-		}
-	})
-}
-
