+++ internal/sidekick/swift/convert_proto_names.go
@@ -1,0 +1,56 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package swift
+
+import (
+	"fmt"
+	"slices"
+	"strings"
+
+	"github.com/googleapis/librarian/internal/sidekick/api"
+)
+
+// protoMessageTypeName resolves the fully-qualified Swift type name generated by SwiftProtobuf
+// for a given api.Message. It constructs the name by walking up the message parent hierarchy,
+// formatting type names in PascalCase, prepending the proto package prefix, and qualifying
+// the path with ModulePath if it is not empty.
+func (c *codec) protoMessageTypeName(m *api.Message) string {
+	var path []string
+	curr := m
+	for curr != nil {
+		path = append(path, pascalCase(curr.Name))
+		curr = curr.Parent
+	}
+	slices.Reverse(path)
+	fullName := fmt.Sprintf("%s%s", ProtoPackagePrefix(m.Package), strings.Join(path, "."))
+	if c.ModulePath != "" {
+		return fmt.Sprintf("%s.%s", c.ModulePath, fullName)
+	}
+	return fullName
+}
+
+// protoEnumTypeName resolves the fully-qualified Swift type name generated by SwiftProtobuf
+// for a given api.Enum. For top-level enums, it formats the package prefix and enum name.
+// For nested enums, it resolves the parent message's proto type name and appends the PascalCase enum name.
+func (c *codec) protoEnumTypeName(e *api.Enum) string {
+	if e.Parent == nil {
+		fullName := fmt.Sprintf("%s%s", ProtoPackagePrefix(e.Package), pascalCase(e.Name))
+		if c.ModulePath != "" {
+			return fmt.Sprintf("%s.%s", c.ModulePath, fullName)
+		}
+		return fullName
+	}
+	return fmt.Sprintf("%s.%s", c.protoMessageTypeName(e.Parent), pascalCase(e.Name))
+}
